Bottom sheet temperature rounded to int

+++ app/src/main/java/gustavo/projects/pinweathermap/domain/model/LocationWeather.kt
@@ -4,17 +4,17 @@ data class LocationWeather(
     val name: String,
     val latitude: Double,
     val longitude: Double,
-    val temp: Double, //°C
-    val feelsLike: Double, //°C
-    val tempMin: Double, //°C
-    val tempMax: Double, //°C
-    val humidity: Double, //%
-    val pressure: Double, //hPa
+    val temp: Int, //°C
+    val feelsLike: Int, //°C
+    val tempMin: Int, //°C
+    val tempMax: Int, //°C
+    val humidity: Int, //%
+    val pressure: Int, //hPa
     val unixTime: Long, // Unix time
     val sunRiseTime: Long, // Unix Time
     val sunSetTime: Long, //Unix Time
     val windSpeed: Double, // Meter/Second
     val main: String,
     val description: String,
-    val clouds: Double //%,
+    val clouds: Int //%,
 )
