+++ app/src/main/java/gustavo/projects/pinweathermap/network/response/CloudsDto.kt
@@ -1,5 +1,5 @@
 package gustavo.projects.pinweathermap.network.response
 
 data class CloudsDto(
-    val all: Double
+    val all: Int
 )
